+++ src/domain/QuestionList/QuestionListCard/QuestionListCardLarge.tsx
@@ -2,7 +2,7 @@ import { MoreOutlined } from '@ant-design/icons';
 import { Col, Dropdown, Progress, Row, Space } from 'antd';
 import React from 'react';
 
-import { difficultyPercentageToColor } from '../../../services/mathUtils';
+import { difficultyPercentageToColor } from '../../../helper/mathUtils';
 import AverageDifficultyContent from '../AverageDifficultyContent';
 import styles from './QuestionListCard.module.scss';
 import { QuestionListCardLargeProps } from './QuestionListCardLargeProps';
