+++ src/helper/stringUtils.ts
