+++ src/helper/table/FilterDropdown.module.scss
@@ -1,0 +1,12 @@
+.padded {
+    padding: 8px;
+}
+
+.filter {
+    margin-bottom: 8px;
+    display: block;
+}
+
+.small {
+    width: 90;
+}
