--- src/services/table/FilterDropdown.module.scss
@@ -1,12 +1,0 @@
-.padded {
-    padding: 8px;
-}
-
-.filter {
-    margin-bottom: 8px;
-    display: block;
-}
-
-.small {
-    width: 90;
-}
